RWA-1062: Prevent inconsistent / duplicate lab orders
This corresponds to recent changes to mohbilling, where mohbilling handles creation of bills when lab orders are saved, and so this functionality is removed to ensure duplicate bills are not issued for the same lab test.  This also adds several other fixes, including only creating a lab appointment for the patient if they do not already have one, and excluding discontinue orders, expired orders, and cancelled orders from the lab lists.